requirements.txt
torch==2.1.0
torchvision==0.16.0
timm==0.9.0
numpy==1.26.1
pandas==2.1.1
scikit-learn==1.4.2
umap-learn==0.5.4
matplotlib==3.8.1
seaborn==0.13.2
pyyaml==6.0.2
pytest==8.3.1
Pillow==10.0.1
